refactor(test): discount 패키지 세분화
