+++ src/test/java/christmas/domain/discount/DiscountManagerTest.java
@@ -1,0 +1,5 @@
+package christmas.domain.discount;
+
+public class DiscountManagerTest {
+
+}
