+++ src/test/java/christmas/domain/discount/strategies/WeekdayDiscountTest.java
@@ -1,4 +1,4 @@
-package christmas.domain.discount;
+package christmas.domain.discount.strategies;
 
 import static christmas.domain.discount.constants.DiscountInfo.WEEKDAY_DISCOUNT_PRICE;
 import static christmas.domain.discount.constants.DiscountInfo.WEEKDAY_DISCOUNT_STRING;
